// src/autorizacao/dtos/create-autorizacao.dto.ts
export class CreateAutorizacaoDto {
  userIdAut: number;
  userIdAlAut: number;
  motivoAut: string;
  dataInicio: string;
  dataFinal: string;
  horaInicio: string;
  horaFinal: string;
  seg: string;
  ter: string;
  qua: string;
  qui: string;
  sex: string;
  sab: string;
  dom: string;
  obsAut: string;
}